pyproject.toml
[tool.poetry]
name = "test-flask"
version = "0.1.2"
description = "flask app"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
  { include = "test_flask" },
]

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
flake8 = "^7.1.1"
jinja2 = "^3.1.4"
flask-wtf = "^1.2.2"
python-dotenv = "^1.0.1"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
